requirements.txt
asgiref==3.6.0
asn1crypto==1.5.1
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
certvalidator==0.11.1
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
endesive==2.0.15
icpyedu==0.0.1
idna==3.4
lxml==4.9.2
Markdown==3.4.1
oscrypto==1.3.0
paramiko==3.0.0
Pillow==9.4.0
pycparser==2.21
PyKCS11==1.5.11
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
django-cors-headers==3.14.0
python-decouple==3.7
whitenoise==6.4.0